+++ policy-engine/src/main/java/com/walmartlabs/concord/policyengine/QueueRule.java
@@ -28,24 +28,31 @@ import java.io.Serializable;
 
 public class QueueRule implements Serializable {
 
+    public static QueueRule empty() {
+        return new QueueRule(null, null, null, null, null, null);
+    }
+
     private final ConcurrentProcessRule concurrent;
     private final QueueProcessRule process;
     private final QueueProcessRule processPerOrg;
     private final QueueProcessRule processPerProject;
     private final ForkDepthRule forkDepthRule;
+    private final ProcessTimeoutRule processTimeoutRule;
 
     @JsonCreator
     public QueueRule(@JsonProperty("concurrent") ConcurrentProcessRule concurrent,
                      @JsonProperty("process") QueueProcessRule process,
                      @JsonProperty("processPerOrg") QueueProcessRule processPerOrg,
                      @JsonProperty("processPerProject") QueueProcessRule processPerProject,
-                     @JsonProperty("forkDepth") ForkDepthRule forkDepthRule) {
+                     @JsonProperty("forkDepth") ForkDepthRule forkDepthRule,
+                     @JsonProperty("processTimeout")  ProcessTimeoutRule processTimeoutRule) {
 
         this.concurrent = concurrent;
         this.process = process;
         this.processPerOrg = processPerOrg;
         this.processPerProject = processPerProject;
         this.forkDepthRule = forkDepthRule;
+        this.processTimeoutRule = processTimeoutRule;
     }
 
     public ConcurrentProcessRule getConcurrent() {
@@ -68,6 +75,10 @@ public class QueueRule implements Serializable {
         return forkDepthRule;
     }
 
+    public ProcessTimeoutRule getProcessTimeoutRule() {
+        return processTimeoutRule;
+    }
+
     @Override
     public String toString() {
         return new ToStringBuilder(this, Utils.NotNullToStringStyle.NOT_NULL_STYLE)
@@ -75,7 +86,8 @@ public class QueueRule implements Serializable {
                 .append("process", process)
                 .append("processPerOrg", processPerOrg)
                 .append("processPerProject", processPerProject)
-                .append("forkDepthRule", forkDepthRule)
+                .append("forkDepth", forkDepthRule)
+                .append("processTimeout", processTimeoutRule)
                 .toString();
     }
 }